move device functional to tensor function file
