+++ rising/transforms/functional/__init__.py
@@ -15,7 +15,6 @@ Implementations include:
 """
 
 from rising.transforms.functional.crop import *
-from rising.transforms.functional.device import *
 from rising.transforms.functional.intensity import *
 from rising.transforms.functional.spatial import *
 from rising.transforms.functional.tensor import *
