+++ tests/transforms/functional/test_device.py
@@ -1,7 +1,7 @@
 import torch
 import unittest
 
-from rising.transforms.functional.device import *
+from rising.transforms.functional.tensor import *
 
 
 class TestSpatialFunctional(unittest.TestCase):
@@ -11,24 +11,24 @@ class TestSpatialFunctional(unittest.TestCase):
 
     @unittest.skipUnless(torch.cuda.is_available(), "No cuda gpu available")
     def test_to_device_tensor(self):
-        outp = to_device(self.batch_2d, self.device)
+        outp = to_device_dtype(self.batch_2d, device=self.device)
         self.assertEqual(outp.device, torch.device(self.device))
 
     @unittest.skipUnless(torch.cuda.is_available(), "No cuda gpu available")
     def test_to_device_mapping(self):
-        outp = to_device({"a": self.batch_2d}, self.device)
+        outp = to_device_dtype({"a": self.batch_2d}, device=self.device)
         self.assertEqual(outp["a"].device, torch.device(self.device))
         self.assertIsInstance(outp, dict)
 
     @unittest.skipUnless(torch.cuda.is_available(), "No cuda gpu available")
     def test_to_device_iterable(self):
-        outp = to_device((self.batch_2d, ), self.device)
+        outp = to_device_dtype((self.batch_2d, ), device=self.device)
         self.assertEqual(outp[0].device, torch.device(self.device))
         self.assertIsInstance(outp, tuple)
 
     @unittest.skipUnless(torch.cuda.is_available(), "No cuda gpu available")
     def test_to_device_str(self):
-        outp = to_device("test", self.device)
+        outp = to_device_dtype("test", device=self.device)
         self.assertIsInstance(outp, str)
